forgot to add extra param in query

+++ initdb.sql
@@ -3,5 +3,5 @@ create table test_table
     id serial primary key,
     col_a integer,
     col_b integer,
-    test_value numeric
+    test_value integer
 );
